@@ -1,30 +1,30 @@
 {
-  "name": "client",
-  "version": "1.5.0",
+  "name": "my_portfolio",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.1",
     "@tinymce/tinymce-react": "^3.8.1",
-    "firebase": "^7.24.0",
+    "firebase": "^8.1.1",
     "jsonwebtoken": "^8.5.1",
-    "jwt-decode": "^3.0.0",
+    "jwt-decode": "^3.1.2",
     "moment": "^2.29.1",
-    "next": "9.4.4",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
+    "next": "10.0.2",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
     "react-html-parser": "^2.0.2",
-    "react-modal": "^3.11.2",
+    "react-modal": "^3.12.1",
     "react-responsive-carousel": "^3.2.10",
-    "sitemap": "^6.3.2",
-    "sweetalert2": "^9.17.2"
+    "sitemap": "^6.3.3",
+    "sweetalert2": "^10.10.1"
   },
   "devDependencies": {
     "postcss-preset-env": "^6.7.0",
-    "tailwindcss": "^1.9.6"
+    "tailwindcss": "^2.0.1"
   }
 }
