@@ -1,6 +1,6 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 
 export default (req, res) => {
-  res.statusCode = 200;
-  res.json({ name: "John Doe" });
-};
+  res.statusCode = 200
+  res.json({ name: 'John Doe' })
+}
